+++ backend/modules/algoProviders/integratedAlgoProvider/integratedAlgoProvider.py
@@ -1,4 +1,6 @@
 import os
+from modules.algoProviders.AlgoProvider import AlgoProvider
+from modules.algoProviders.AlgoProviderException import AlgoProviderException
 
 
 def _get_algorithm_python(algorithm_name):
@@ -31,55 +33,68 @@ def _get_algorithm_python(algorithm_name):
     return (algorithm_name, algorithm_python)
 
 
-def get_algorithms():
-    """Get all algorithms that DebiAI can provide
+class IntegratedAlgoProvider(AlgoProvider):
+    ### Integrated AlgoProvider
+    # Used to expose the algorithms that are integrated
+    # directly in DebiAI
+    def __init__(self):
+        self.url = "/app/algo-provider"
+        self.name = "Integrated Algo-provider"
+        self.alive = True
 
-    Returns:
-        list: List of algorithms
-    """
+    def is_alive(self):
+        return True
 
-    # List the .py files if the algorithms folder
-    algorithm_files = []
-    for file in os.listdir(os.path.dirname(__file__) + "/algorithms"):
-        if file.endswith(".py") and file != "__init__.py":
-            algorithm_files.append(file[:-3])
+    def get_algorithms(self):
+        """Get all algorithms that DebiAI can provide
+        Returns:
+            list: List of algorithms
+        """
 
-    # Import the algorithms
-    algorithms_python = []
-    for file in algorithm_files:
-        print("Importing " + file)
+        # List the .py files if the algorithms folder
+        algorithm_files = []
+        for file in os.listdir(os.path.dirname(__file__) + "/algorithms"):
+            if file.endswith(".py") and file != "__init__.py":
+                algorithm_files.append(file[:-3])
+
+        # Import the algorithms
+        algorithms_python = []
+        for file in algorithm_files:
+            print("   Importing " + file)
+            try:
+                algorithms_python.append(_get_algorithm_python(file))
+            except ModuleNotFoundError as e:
+                print("Error importing " + file)
+                print(e)
+
+        # Get the algorithms (call the get_algorithm_details() function)
+        algorithms = []
+        for algorithm in algorithms_python:
+            algorithm_details = algorithm[1].get_algorithm_details()
+            # Add the id as the file name
+            algorithm_details["id"] = algorithm[0]
+            algorithms.append(algorithm_details)
+
+        return algorithms
+
+    def use_algorithm(self, algorithm_id, data):
         try:
-            algorithms_python.append(_get_algorithm_python(file))
-        except ModuleNotFoundError as e:
-            print("Error importing " + file)
+            print("Using integrated algo-provider")
+            print("Using algorithm: " + algorithm_id)
+            algorithm = _get_algorithm_python(algorithm_id)
+
+            # Use the algorithm
+            outputs = algorithm[1].use_algorithm(data["inputs"])
+
+            return outputs
+
+        except TypeError as e:
+            print("The integrated algo-provider returned an error")
             print(e)
-
-    # Get the algorithms (call the get_algorithm_details() function)
-    algorithms = []
-    for algorithm in algorithms_python:
-        algorithm_details = algorithm[1].get_algorithm_details()
-        # Add the id as the file name
-        algorithm_details["id"] = algorithm[0]
-        algorithms.append(algorithm_details)
-
-    return algorithms
-
-
-def use_algorithm(algorithmId, data):
-    """Use an algorithm
-
-    Args:
-        algorithmId (string): ID of the algorithm
-        data (list): inputs
-
-    Returns:
-        list: outputs
-    """
-
-    # Get the algorithm
-    algorithm = _get_algorithm_python(algorithmId)
-
-    # Use the algorithm
-    outputs = algorithm[1].use_algorithm(data["inputs"])
-
-    return outputs
+            raise AlgoProviderException(
+                algorithm_id + " returned an error: " + str(e), 400
+            )
+        except Exception as e:
+            print("The integrated algo-provider returned an error")
+            print(e)
+            raise AlgoProviderException("AlgoProvider internal server error", 500)
